change id type from Long to Integer

+++ scrapper/src/main/java/edu/java/domain/entities/Link.java
@@ -2,4 +2,4 @@ package edu.java.domain.entities;
 
 import java.net.URI;
 
-public record Link(Long id, URI uri) { }
+public record Link(Integer id, URI uri) { }
